pyproject.toml
[project]
name = "order-inventory-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.1",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.128.0",
    "greenlet>=3.3.0",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pwdlib>=0.3.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
